pyproject.toml
[tool.poetry]
name = "qaseio"
version = "0.1.0"
description = ""
license = "Apache-2.0"
authors = ["Parviz Khavari <csctclan@gmail.com>"]
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3"
]

[tool.poetry.dependencies]
python = "^3.6"
python-apitist = "^0.5.0"
attrs = "^19.3.0"
cattrs = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^4.6"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
